refactor: rename util modules
