+++ src/netrics/measurement/hops_traceroute.py
@@ -6,7 +6,7 @@ import typing
 from schema import Optional
 
 from netrics import task
-from netrics.util.iteration import sequence
+from netrics.util.iterutils import sequence
 
 from .common import (
     default,
