+++ src/netrics/util/iterutils.py
