+++ src/netrics/util/lazy_validation.py
